+++ assistantSkillsMerger.js
@@ -4,7 +4,7 @@ function combineWatsonSkills(file11, file22) {
     var combinedJSON = mergeIntentsAndEntities(file1, file2);
     // var indexOfAppend = searchNodeByTitle(file1, "Anything else");
     var indexOfAppend = findLastDialogNode(file1);
-    combileDialogNodes(file1, indexOfAppend, file2);
+    mergeDialogNodes(file1, indexOfAppend, file2);
     var outputSkillName = updateMergedSkillProp(file1.name, file2.name, "skillname");
     var outputSkillDesc = updateMergedSkillProp(file1.description, file2.description, "description");
     combinedJSON.name = outputSkillName; //Update new skill name before writing to file
@@ -21,21 +21,19 @@ function mergeIntentsAndEntities(file1, file2) {
     var entitiesFile2 = file2.entities;
     var tempHashMap = {};
     //Add skills and entities from Skill2 to skill1
-    //This currently doesn't cater for clashes
+
     var intentsInFile1 = file1.intents; //Arrays on File1 intents
     var entitiesInFile1 = file1.entities; //Arrays on File1 intents
 
-    for (intentObject in intentsInFile1) { //Copies each intents into temp hash
+    for (intentObject in intentsInFile1) { //Copies each intents into tempHashMap
         var intentName = intentsInFile1[intentObject].intent;
-        // console.log(intentName);
         tempHashMap[intentName.toLowerCase()] = intentName;
     }
+
     for (intentObject in intentsArr2) {
         var intentName = intentsArr2[intentObject].intent;
-        if (!tempHashMap[intentName.toLowerCase()]) { //Avoid duplicate intents, check Cases
-            // console.log(intentName+ " doesnt exist and added!")
+        if (!tempHashMap[intentName.toLowerCase()]) //Avoid duplicate intents, check Cases
             file1.intents.push(intentsArr2[intentObject]); //adds intents from Skills2 to Skill1
-        }
     }
 
     var tempEntityHashMap = {};
@@ -45,23 +43,12 @@ function mergeIntentsAndEntities(file1, file2) {
     }
 
     for (entityObject in entitiesFile2) {
-        var entityName = entitiesFile2[entityObject].entity;A
+        var entityName = entitiesFile2[entityObject].entity;
         if (!tempEntityHashMap[entityName.toLowerCase()])
             file1.entities.push(entitiesFile2[entityObject]); //adds entities from Skills2 to Skill
     }
     // console.log(file1);
     return file1;
-}
-
-function searchNodeByTitle(jsonToSearch, searchValue) {
-    //Expected inputs are json, returns nodePos
-    var dialogueNodeFile = jsonToSearch.dialog_nodes;
-    for (eachNode in dialogueNodeFile) {
-        var currentNode = dialogueNodeFile[eachNode];
-        if (currentNode.title === searchValue)
-            return eachNode; //Position of lastNode
-    }
-    return -1;
 }
 
 function findFirstDialogNodeLoc(jsonToSearch) {
@@ -86,7 +73,7 @@ function findLastDialogNode(jsonToSearch) {
         if (!(jsonToSearch.dialog_nodes[i].parent)) { //Add all root nodes to Hash
             var prev = jsonToSearch.dialog_nodes[i].previous_sibling; //Add all previous
             // console.log("Previous is "+ prev);
-            hashMapReferredTo[prev] =1; //push all previous item to hashmap
+            hashMapReferredTo[prev] = 1; //push all previous item to hashmap
         }
     }
     //Checking for Items not referred to
@@ -102,76 +89,13 @@ function findLastDialogNode(jsonToSearch) {
     }
     return -1;
 }
-function deleteNode(jsonToDeleteFrom, nodeIndexToDelete) {
-    if (!nodeIndexToDelete) {
-        jsonToDeleteFrom.dialog_nodes.pop();
-        // console.log(jsonToDeleteFrom.dialog_nodes);
-    } else {
-        // console.log(jsonToDeleteFrom.dialog_nodes);
-        var removedNode = jsonToDeleteFrom.dialog_nodes.splice(nodeIndexToDelete, 1);
-        // console.log(removedNode);
-        // console.log(removedNode.title + " was removed from DialogNodes")
-    }
-}
 
-function combileDialogNodes(file1, indexofAppend, file2) {
-    console.log("\n* combining dialog nodes *")
-    //indexOfAppend is the index of last element
-    //Expected inputs are json
-    var lastItemFromFile1 = file1.dialog_nodes[indexofAppend].dialog_node; //Temp Storage
-    console.log("Last Node in file1-> " + lastItemFromFile1);
-
-    //Logs all current dialogueNodes in HashMap
-    var tempHashMapDiagNodes = {};
-    var diaNodesInFile1 = file1.dialog_nodes; //Arrays on File1 dialogNodes
-    for (node in diaNodesInFile1) //Keep track of all dialog nodes in file1
-        tempHashMapDiagNodes[diaNodesInFile1[node].dialog_node] = diaNodesInFile1[node].dialog_node;
-    // console.log(tempHashMapDiagNodes);
-
-    var firstNodeInFile2 = findFirstDialogNodeLoc(file2); //gets location of first Node in file2 for MERGE
-
-    var renamedNodesMap = {};
-    var dialogNodes2 = file2.dialog_nodes; //SecondObject
-    for (dialogNode in dialogNodes2) { //Iterating File2
-        var currentNode = dialogNodes2[dialogNode];
-        //Update CurrentNode Parent and/or siblings
-        if (tempHashMapDiagNodes[currentNode.parent]) {
-
-            if (renamedNodesMap[currentNode.parent]) //If already renamed use, otherwise create
-                currentNode.parent = renamedNodesMap[currentNode.parent];
-            else
-                currentNode.parent = currentNode.parent + '_1';
-            // console.log("Yes " + currentNode.dialog_node + " has  parent " + currentNode.parent)
-        }
-        else if (tempHashMapDiagNodes[currentNode.previous_sibling]) {
-
-            if (renamedNodesMap[currentNode.previous_sibling]) //If already renamed use, otherwise create
-                currentNode.previous_sibling = renamedNodesMap[currentNode.previous_sibling];
-            else
-                currentNode.previous_sibling = currentNode.previous_sibling + '_1';
-
-            // console.log("Yes " + currentNode.dialog_node + " has  previous node " + currentNode.previous_sibling);
-        }
-        if (tempHashMapDiagNodes[currentNode.dialog_node]) {
-            console.log("\nFound " + currentNode.title + " to be conflicting");
-            //If current Node in file2 is a conflict, update it and push to file1
-            // console.log("Yes "+ currentNode.dialog_node + " is a conflict!");
-            if (renamedNodesMap[currentNode.dialog_node]) //If already renamed use, otherwise create
-                currentNode.dialog_node = renamedNodesMap[currentNode.dialog_node];
-            else
-                currentNode.dialog_node = currentNode.dialog_node + '_1';
-
-            console.log("After conflict, NODEID renamed to " + currentNode.dialog_node);
-        }
-        if (dialogNode == firstNodeInFile2)
-            dialogNodes2[dialogNode].previous_sibling = lastItemFromFile1;     //i.e. Last of File1 + first of File2
-        file1.dialog_nodes.push(currentNode);
-
-    }
-
-    // console.log(renamedNodesMap);
-    // file1.name = file1.name + "-" + file2.name;
-    return file1;
+function getUniqueNodeID(listOfKnownNodes, incoming) {
+    var renamed = incoming;
+    while (listOfKnownNodes[renamed])
+        renamed = renamed + '_1';
+    console.log(incoming + " was renamed to " + renamed);
+    return renamed;
 }
 
 function writeToFile(jsonData, outputFileName, fileType) {
@@ -232,9 +156,54 @@ function runOutputFileReport(combinedJSON) {
         console.log("*** Please Note: Output File will be generated, however will not upload in Watson");
     }
 }
-//Test Module
-// var file1 = require('./sample-skills-files/skill-SkillOne');
-// var file2 = require('./sample-skills-files/skill-SkillTwo');
-// combineWatsonSkills(file1, file2);
+
+function mergeDialogNodes(file1, locFile1LastItem, file2) {
+    console.log("\n* combining dialog nodes *");
+    //indexOfAppend is the index of last element. Expected inputs are json
+    var file1LastNodeID = file1.dialog_nodes[locFile1LastItem].dialog_node;
+    var file2FirstNodeID = findFirstDialogNodeLoc(file2); //gets location of first Node in file2 for MERGE
+
+    var file1DialogNodesArr = file1.dialog_nodes; // File1's dialogNodes Array
+    var tempHashMapDiagNodes = {}; //Container for File1's dialogID
+    for (var i in file1DialogNodesArr) //Stores each dialogID of firstFile in a tempHashMap
+        tempHashMapDiagNodes[file1DialogNodesArr[i].dialog_node] = file1DialogNodesArr[i].dialog_node;
+    console.log("displaying all Nodes: \n" + JSON.stringify(tempHashMapDiagNodes) + "\n"); //DELETE
+
+    var renamedNodesMap = {}; //
+    var dialogNodes2 = file2.dialog_nodes; //File2's dialogNodes Array
+
+    for (var i in dialogNodes2) { //Adding DialogNodes from file2 to file1
+        var currentNode = dialogNodes2[i];
+
+        //Conflicts resolution
+        if (!renamedNodesMap[dialogNodes2]) { // if Item in file2 is not already Renamed and is a Conflict, Remember it by storing to Renamed
+            if (tempHashMapDiagNodes[currentNode.dialog_node])
+                renamedNodesMap[currentNode.dialog_node] = getUniqueNodeID(tempHashMapDiagNodes, currentNode.dialog_node);
+
+            if (tempHashMapDiagNodes[currentNode.parent])
+                renamedNodesMap[currentNode.parent] = getUniqueNodeID(tempHashMapDiagNodes, currentNode.parent);
+
+            if (tempHashMapDiagNodes[currentNode.previous_sibling])
+                renamedNodesMap[currentNode.previous_sibling] = getUniqueNodeID(tempHashMapDiagNodes, currentNode.previous_sibling);
+        }
+        else { // if Item in file2 is already renamed, The use Renamed Value;
+            if (tempHashMapDiagNodes[currentNode.dialog_node]) {
+                renamedNodesMap[currentNode.dialog_node] = renamedNodesMap[currentNode.dialog_node];
+            }
+            if (tempHashMapDiagNodes[currentNode.parent])
+                renamedNodesMap[currentNode.parent] = renamedNodesMap[currentNode.parent];
+            if (tempHashMapDiagNodes[currentNode.previous_sibling])
+                renamedNodesMap[currentNode.previous_sibling] = renamedNodesMap[currentNode.previous_sibling];
+        }
+
+        //The merging of lastNode from File1 and firstNode from File2
+        if (i == file2FirstNodeID)
+            currentNode.previous_sibling = file1LastNodeID;     //i.e. Last of File1 + first of File2
+
+        if (!tempHashMapDiagNodes[currentNode.dialog_node])
+            file1.dialog_nodes.push(currentNode); //Adds all nodes file2 to file1
+    }
+    return file1;
+}
 
 module.exports = combineWatsonSkills;